Revert "Add WrapperAware trait"
This reverts commit 241abc4317eec701211b7a88a17a1b610c366dfe.
