+++ src/PersistentModel.php
@@ -10,7 +10,6 @@ use Rubix\ML\Serializers\Serializer;
 use Rubix\ML\AnomalyDetectors\Scoring;
 use Rubix\ML\Exceptions\InvalidArgumentException;
 use Rubix\ML\Exceptions\RuntimeException;
-use Rubix\ML\Traits\WrapperAware;
 
 /**
  * Persistent Model
@@ -24,7 +23,12 @@ use Rubix\ML\Traits\WrapperAware;
  */
 class PersistentModel implements Wrapper, Learner, Probabilistic, Scoring
 {
-    use WrapperAware;
+    /**
+     * The persistable base learner.
+     *
+     * @var Learner
+     */
+    protected \Rubix\ML\Learner $base;
 
     /**
      * The persister used to interface with the storage layer.
@@ -81,6 +85,30 @@ class PersistentModel implements Wrapper, Learner, Probabilistic, Scoring
     }
 
     /**
+     * Return the estimator type.
+     *
+     * @internal
+     *
+     * @return EstimatorType
+     */
+    public function type() : EstimatorType
+    {
+        return $this->base->type();
+    }
+
+    /**
+     * Return the data types that the estimator is compatible with.
+     *
+     * @internal
+     *
+     * @return list<\Rubix\ML\DataType>
+     */
+    public function compatibility() : array
+    {
+        return $this->base->compatibility();
+    }
+
+    /**
      * Return the settings of the hyper-parameters in an associative array.
      *
      * @internal
@@ -107,6 +135,16 @@ class PersistentModel implements Wrapper, Learner, Probabilistic, Scoring
     }
 
     /**
+     * Return the base estimator instance.
+     *
+     * @return Estimator
+     */
+    public function base() : Estimator
+    {
+        return $this->base;
+    }
+
+    /**
      * Save the model to storage.
      */
     public function save() : void
@@ -128,6 +166,17 @@ class PersistentModel implements Wrapper, Learner, Probabilistic, Scoring
     public function train(Dataset $dataset) : void
     {
         $this->base->train($dataset);
+    }
+
+    /**
+     * Make a prediction on a given sample dataset.
+     *
+     * @param Dataset $dataset
+     * @return mixed[]
+     */
+    public function predict(Dataset $dataset) : array
+    {
+        return $this->base->predict($dataset);
     }
 
     /**
